move bn api class to lib/bottlenose
