+++ app/controllers/api/professor/courses_controller.rb
@@ -25,24 +25,7 @@ module Api
       end
 
       def sync
-        got = bottlenose_get("/api/courses/#{@course.bottlenose_id}/registrations")
-        got.each do |sec_id, sec_obj|
-          sec = @course.sections.find_or_initialize_by(bottlenose_id: sec_id)
-          sec.title = "#{sec_obj['type']} - #{sec_obj['meeting_time']}"
-          sec.save!
-          sec_obj['students'].each do |student|
-            user = sync_user(student)
-            reg = sec.student_registrations.find_or_initialize_by(user: user)
-            reg.save!
-          end
-          sec_obj['staff'].each do |staff|
-            user = sync_user(staff['user'])
-            reg = sec.staff_registrations.find_or_initialize_by(user: user)
-            reg.ta = staff['ta']
-            reg.save!
-          end
-          # TODO profs
-        end
+        bottlenose_api.sync_course(@course)
         render json: { synced: true }
       rescue Bottlenose::UnauthorizedError => e
         render json: {
@@ -54,18 +37,6 @@ module Api
           synced: false,
           reason: e.message
         }
-      end
-
-      private
-
-      def sync_user(u)
-        user = User.where(username: u['username']).first_or_initialize
-        user.display_name = u['display_name']
-        user.nuid = u['nuid']
-        user.email = u['email']
-        user.image_url = u['image_url']
-        user.save!
-        user
       end
     end
   end
